requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
decorator==5.0.9
idna==2.10
iniconfig==1.1.1
jsonpath-rw==1.4.0
packaging==21.0
pluggy==0.13.1
ply==3.11
py==1.10.0
PyEmail==0.0.1
pyparsing==2.4.7
pytest==6.2.4
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.6
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
